+++ src/main/java/courgette/integration/slack/SlackPublisher.java
@@ -1,11 +1,9 @@
 package courgette.integration.slack;
 
+import courgette.runtime.event.CourgetteEventHolder;
 import courgette.runtime.CourgetteProperties;
 import courgette.runtime.CourgettePublisher;
-import courgette.runtime.CourgetteRunResult;
-import courgette.runtime.CourgetteRunnerInfo;
 import courgette.runtime.CourgetteSlackOptions;
-import courgette.runtime.event.CourgetteEvent;
 import courgette.runtime.event.EventPublisher;
 import courgette.runtime.event.EventSubscriberCreator;
 
@@ -14,13 +12,9 @@ import java.util.Optional;
 
 public class SlackPublisher implements CourgettePublisher {
 
-    private final CourgetteProperties courgetteProperties;
-
     private Optional<EventPublisher> eventPublisher = Optional.empty();
 
     public SlackPublisher(CourgetteProperties courgetteProperties) {
-
-        this.courgetteProperties = courgetteProperties;
 
         if (courgetteProperties.publishEventsToSlack()) {
             final CourgetteSlackOptions slackOptions = courgetteProperties.slackOptions();
@@ -34,17 +28,7 @@ public class SlackPublisher implements CourgettePublisher {
     }
 
     @Override
-    public void publish(CourgetteEvent event) {
-        eventPublisher.ifPresent(p -> p.publishEvent(event, courgetteProperties, null, null));
-    }
-
-    @Override
-    public void publish(CourgetteEvent event, CourgetteRunResult result) {
-        eventPublisher.ifPresent(p -> p.publishEvent(event, courgetteProperties, null, result));
-    }
-
-    @Override
-    public void publish(CourgetteEvent event, CourgetteRunnerInfo runnerInfo, CourgetteRunResult result) {
-        eventPublisher.ifPresent(p -> p.publishEvent(event, courgetteProperties, runnerInfo, result));
+    public void publish(CourgetteEventHolder eventHolder) {
+        eventPublisher.ifPresent(p -> p.publishEvent(eventHolder));
     }
 }